refactoring, testing etc. Ran into weird issue, I do not think the edge polling example from uasyncio even can work...
