+++ scripts/gps_test.py
@@ -1,13 +1,17 @@
 import rtb
 
 import logging
-#logging.basicConfig(logging.DEBUG)
-from uasyncio.core import get_event_loop
+logging.basicConfig(logging.DEBUG)
+from uasyncio.core import get_event_loop,sleep
 
 from rtb.gps import instance as gps
 rtb.pwr.GPS_VCC.status()
 
-gps.start()
+
+get_event_loop().create_task(gps.start())
+
+get_event_loop().create_task(rtb.heartbeat(1))
+
 
 loop = get_event_loop()
 loop.run_forever()
