+++ stmhal/boards/RUUVITRACKER_C3/copy_to_board/rtb/__init__.py
@@ -1,6 +1,8 @@
 import pyb
 # Muck the uasyncio default eventloop as soon as rtb is imported
 from . import eventloop
+from uasyncio.core import sleep
+
 from .powerdomains import powermanager_singleton
 from . import powerdomains_config as pwr
 
@@ -18,3 +20,25 @@ GSM_DGB_UART_N = None
 #///   - `UART(6)` is on `YA`: `(TX, RX) = (Y1, Y2) = (PC6, PC7)`
 #///   - `UART(3)` is on `YB`: `(TX, RX) = (Y9, Y10) = (PB10, PB11)`
 #///   - `UART(2)` is on: `(TX, RX) = (X3, X4) = (PA2, PA3)`
+
+# This is a coroutine, we do not use the decorator to indicate that due to resource constrainst of pyboard
+def heartbeat(ledno=1):
+    led = pyb.LED(ledno)
+    sleep1 = 2000
+    # It seems that when RTC is working we cannot really do subsecond sleeps, and if it's not working we cannot do concurrent sleeps...
+    sleep2 = 100
+    sleep3 = 250
+    while True:
+        yield from sleep(sleep1)
+        led.on()
+        #print("led %s ON" % ledno)
+        yield from sleep(sleep2)
+        led.off()
+        #print("led %s OFF" % ledno)
+        yield from sleep(sleep3)
+        led.on()
+        #print("led %s ON" % ledno)
+        yield from sleep(sleep2)
+        led.off()
+        #print("led %s OFF" % ledno)
+        #print("Looping back")
