+++ stmhal/boards/RUUVITRACKER_C3/copy_to_board/uasyncio/__init__.py
@@ -1,2 +1,2 @@
 # Two purposes 1. make this dis a module 2. if imported directly, import the pybselect eventloop
-from . import pybselect
+from .pybselect import *
